Objective of test cases: Water Bottle
Requirements:
What is the material of a water bottle? Is it a bottle made of glass, plastic, rubber, metal, disposable material, or other material?
Is there any alternate purpose for the water bottle, is it meant only to hold water? Or can we use it with other fluids like tea, coffee, soft drinks, hot chocolate, soups,  wine, cooking oil, vinegar, gasoline, acids, etc.?
Who are the target users of this water bottle? A school-going kid, a housewife, some beverage manufacturing company, an office-goer, a sportsman, or an astronaut in a spaceship?